refactor(config): Move inventory parsing to InventoryConfig
